pyproject.toml
[project]
name = "main"
version = "1.0.0"
description = "Main Controller for Dormitory Project Backend"
readme = "README.md"
authors = [
    { name = "ZeroCore", email = "esetting@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "buptmw>=0.2.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "logto>=0.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "alembic>=1.18.0",
    "gunicorn>=23.0.0",
]
